hw1 2й пункт

+++ src/main/java/hw1_level2/Ready.java
@@ -1,7 +1,7 @@
 package hw1_level2;
 
 public interface Ready {
-    void run(int length);
+    void run(double length);
 
     void jump(int height);
 }
